+++ src/sql/base/query/browser/untitledQueryEditorInput.ts
@@ -13,7 +13,10 @@ import { IResolvedTextEditorModel } from 'vs/editor/common/services/resolverServ
 import { UntitledTextEditorInput } from 'vs/workbench/services/untitled/common/untitledTextEditorInput';
 import { IUntitledTextEditorModel } from 'vs/workbench/services/untitled/common/untitledTextEditorModel';
 import { EncodingMode } from 'vs/workbench/services/textfile/common/textfiles';
-import { EditorInputCapabilities } from 'vs/workbench/common/editor';
+import { GroupIdentifier, ISaveOptions, IEditorInput, EditorInputCapabilities } from 'vs/workbench/common/editor';
+import { FileQueryEditorInput } from 'sql/workbench/contrib/query/browser/fileQueryEditorInput';
+import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
+import { FileEditorInput } from 'vs/workbench/contrib/files/browser/editors/fileEditorInput';
 import { UNTITLED_QUERY_EDITOR_TYPEID } from 'sql/workbench/common/constants';
 import { IUntitledQueryEditorInput } from 'sql/base/query/common/untitledQueryEditorInput';
 
@@ -27,9 +30,10 @@ export class UntitledQueryEditorInput extends QueryEditorInput implements IUntit
 		results: QueryResultsInput,
 		@IConnectionManagementService connectionManagementService: IConnectionManagementService,
 		@IQueryModelService queryModelService: IQueryModelService,
-		@IConfigurationService configurationService: IConfigurationService
+		@IConfigurationService configurationService: IConfigurationService,
+		@IInstantiationService instantiationService: IInstantiationService,
 	) {
-		super(description, text, results, connectionManagementService, queryModelService, configurationService);
+		super(description, text, results, connectionManagementService, queryModelService, configurationService, instantiationService);
 	}
 
 	public override resolve(): Promise<IUntitledTextEditorModel & IResolvedTextEditorModel> {
@@ -42,6 +46,37 @@ export class UntitledQueryEditorInput extends QueryEditorInput implements IUntit
 
 	public get hasAssociatedFilePath(): boolean {
 		return this.text.model.hasAssociatedFilePath;
+	}
+
+	override async save(group: GroupIdentifier, options?: ISaveOptions): Promise<IEditorInput | undefined> {
+		let fileEditorInput = await this.text.save(group, options);
+		return this.createFileQueryEditorInput(fileEditorInput);
+	}
+
+	override async saveAs(group: GroupIdentifier, options?: ISaveOptions): Promise<IEditorInput | undefined> {
+		let fileEditorInput = await this.text.saveAs(group, options);
+		return this.createFileQueryEditorInput(fileEditorInput);
+	}
+
+	private async createFileQueryEditorInput(fileEditorInput: IEditorInput): Promise<IEditorInput> {
+		// Create our own FileQueryEditorInput wrapper here so that the existing state (connection, results, etc) can be transferred from this input to the new file input.
+		try {
+			let newUri = fileEditorInput.resource.toString(true);
+			await this.changeConnectionUri(newUri);
+			this._results.uri = newUri;
+			let newInput = this.instantiationService.createInstance(FileQueryEditorInput, '', (fileEditorInput as FileEditorInput), this.results);
+			newInput.state.setState(this.state);
+			return newInput;
+		}
+		catch (error) {
+			/**
+			 * We are saving over a file that is already open and connected, return the unaltered save editor input directly (to avoid side effects when changing connection).
+			 * This will replace the editor input in the already open window with that of the new one,
+			 * the connection will be undefined and the file appears to be disconnected even if its not, change the connection to fix this.
+			 * also the results shown will be the old results, until run is clicked again.
+			 */
+			return fileEditorInput;
+		}
 	}
 
 	public setMode(mode: string): void {