+++ i18n/ads-language-pack-zh-hans/CHANGELOG.md
@@ -3,3 +3,4 @@ All notable changes to the "ads-language-pack-zh-hans" language pack will be doc
 
 ## [Released]
 * May 21, 2021 - Release for Azure Data Studio 1.29
+* September 2, 2021 - Release for Azure Data Studio 1.33